+++ doc/releaseNotes.md
@@ -561,3 +561,5 @@ This file describes the main feature changes for each readout.exe released versi
 
 ## next version
 - o2-readout-test-fmq-memory: more options.
+- Added protection against unhandled exceptions (e.g. from ReadoutCard).
+- Added protection against unexpected state machine transitions (e.g. ECS sending RESET while STARTING).